Added test canvas, color picker, resize canvas option in MainApp and created new entities, controllers, and the canvas class.
