--- src/main/resources/org/openjfx/canvas.fxml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.StackPane?>
-<?import javafx.scene.control.Button?>
-
-<StackPane xmlns="http://javafx.com/javafx"
-           xmlns:fx="http://javafx.com/fxml"
-           fx:controller="org.openjfx.FXMLController">
-    <children>
-        <Canvas fx:id="canvas" width="800" height="600" />
-        <Button text="Clear" onAction="#clearCanvas" />
-    </children>
-</StackPane>
