+++ src/main/resources/org/openjfx/editor.fxml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.canvas.*?>
+<?import javafx.scene.layout.*?>
+
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.openjfx.Editor">
+   <children>
+      <Canvas height="264.0" layoutY="7.0" width="200.0" />
+   </children>
+</AnchorPane>
